feat(resource): 添加代理节点删除校验功能
- 在ProxyNodeMapper中新增countByInstanceId和countByCustomerId方法
- 在VpsInstanceServiceImpl中添加删除实例前的代理节点校验
- 在MbCustomerServiceImpl中添加删除客户前的代理节点校验
- 在mapper xml文件中实现相应的统计查询sql
- 移除application-druid.yml配置文件中的敏感信息

+++ skyway-member/src/main/java/com/skyway/member/service/impl/MbCustomerServiceImpl.java
@@ -3,10 +3,12 @@ package com.skyway.member.service.impl;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.skyway.common.exception.ServiceException;
 import com.skyway.common.utils.StringUtils;
 import com.skyway.member.domain.MbCustomer;
 import com.skyway.member.mapper.MbCustomerMapper;
 import com.skyway.member.service.IMbCustomerService;
+import com.skyway.resource.mapper.ProxyNodeMapper;
 
 /**
  * 会员客户 服务实现
@@ -18,6 +20,9 @@ public class MbCustomerServiceImpl implements IMbCustomerService {
 
     @Autowired
     private MbCustomerMapper mbCustomerMapper;
+
+    @Autowired
+    private ProxyNodeMapper proxyNodeMapper;
 
     @Override
     public List<MbCustomer> selectList(MbCustomer query) {
@@ -51,11 +56,21 @@ public class MbCustomerServiceImpl implements IMbCustomerService {
 
     @Override
     public int deleteById(Long id) {
+        if (id != null && proxyNodeMapper.countByCustomerId(id) > 0) {
+            throw new ServiceException("该客户下存在代理节点，无法删除，请先移除相关代理节点");
+        }
         return mbCustomerMapper.deleteById(id);
     }
 
     @Override
     public int deleteByIds(Long[] ids) {
+        if (ids != null) {
+            for (Long id : ids) {
+                if (proxyNodeMapper.countByCustomerId(id) > 0) {
+                    throw new ServiceException("该客户下存在代理节点，无法删除，请先移除相关代理节点");
+                }
+            }
+        }
         return mbCustomerMapper.deleteByIds(ids);
     }
